+++ flinkboot-core/src/main/java/io/github/sekelenao/flinkboot/core/internal/resource/FileSystemResource.java
@@ -17,7 +17,7 @@ public class FileSystemResource implements Resource {
     private final String location;
 
     public FileSystemResource(String location) {
-        this.location = Objects.requireNonNull(location);
+        this.location = Objects.requireNonNull(location).replaceFirst("^/+(?=[a-zA-Z]:)", "");
     }
 
     @Override